TODO: after decode, send the image to predict.
