+++ backend/package.json
@@ -22,6 +22,7 @@
     "mongoose": "^5.12.0",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.7",
+    "python-shell": "^2.0.3",
     "redis": "^3.0.2"
   }
 }
