+++ backend/py-files/crop.py
