+++ backend/py-files/flask_project/static/logo.png
